pyproject.toml
[tool.poetry]
name = "swagger-buddy-backend"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.1"
certifi = "^2025.1.31"
alembic = "^1.15.2"
sqlmodel = "^0.0.24"
pendulum = "^3.0.0"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
slowapi = "^0.1.9"
python-multipart = "^0.0.20"
httpx = "^0.28.1"
pyyaml = "^6.0.2"
langchain-openai = "^0.3.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
